feat: complete compose fidelity, security, and standards pass
Faithful docker-compose translation:
- Honor COMPOSE_PROJECT_NAME and COMPOSE_FILE env vars (flag > env > default).
- Forward build.secrets into the build-context tar as id=,src= specs;
  wire cache_to -> cacheto and additional_contexts -> additionalbuildcontexts;
  warn on build.ssh (no libpod REST parameter exists).
- Support .dockerignore negation (!pattern, last-match-wins) and ** recursive
  globs via a backtracking matcher.

Security and correctness:
- Prevent zip-slip in 'cp' container->host extraction: unpack entry-by-entry
  via unpack_in and error on any path that escapes the destination.
- Surface a failed container wait in 'run' instead of masking it as exit 0.
- Serialize mutating commands with a per-project flock(2) advisory lock.

Standards and docs:
- Split every source/test file over the 500-line hard limit into submodules.
- Add .editorconfig; add SAFETY comments to unsafe blocks.
- Extract the libpod /v4.0.0 version prefix into a documented API_PREFIX const.
- README: describe the transport as the native libpod REST API; fix stale pin.
