+++ internal/engine/query.rs
@@ -8,7 +8,7 @@ use crate::libpod::types::exec::{
 	ExecCreateConfig, ExecCreateResponse, ExecInspect, ExecStartConfig,
 };
 use crate::libpod::types::image::ImageInspect;
-use crate::libpod::{urlencoded, LogOutput};
+use crate::libpod::{urlencoded, LogOutput, API_PREFIX};
 
 use super::Engine;
 
@@ -18,7 +18,7 @@ impl Engine {
 		let label = format!("podup.project={}", self.project);
 		let filters = serde_json::json!({ "label": [label] });
 		let path = format!(
-			"/v4.0.0/libpod/containers/json?all=true&filters={}",
+			"{API_PREFIX}/containers/json?all=true&filters={}",
 			urlencoded(&filters.to_string()),
 		);
 
@@ -90,7 +90,7 @@ impl Engine {
 					let client = &self.client;
 					async move {
 						let path = format!(
-							"/v4.0.0/libpod/containers/{}/logs?stdout=true&stderr=true&follow=true",
+							"{API_PREFIX}/containers/{}/logs?stdout=true&stderr=true&follow=true",
 							urlencoded(&container_name),
 						);
 						let resp = match client.get_stream(&path).await {
@@ -123,7 +123,7 @@ impl Engine {
 		} else {
 			for (container_name, is_tty) in targets {
 				let path = format!(
-					"/v4.0.0/libpod/containers/{}/logs?stdout=true&stderr=true&follow={}",
+					"{API_PREFIX}/containers/{}/logs?stdout=true&stderr=true&follow={}",
 					urlencoded(&container_name),
 					follow,
 				);
@@ -174,7 +174,7 @@ impl Engine {
 			..Default::default()
 		};
 		let create_path = format!(
-			"/v4.0.0/libpod/containers/{}/exec",
+			"{API_PREFIX}/containers/{}/exec",
 			urlencoded(&container_name),
 		);
 		let resp: ExecCreateResponse = self
@@ -188,7 +188,7 @@ impl Engine {
 			detach: false,
 			tty: false,
 		};
-		let start_path = format!("/v4.0.0/libpod/exec/{}/start", urlencoded(&exec_id));
+		let start_path = format!("{API_PREFIX}/exec/{}/start", urlencoded(&exec_id));
 		let start_resp = self
 			.client
 			.post_json_stream(&start_path, &start_cfg)
@@ -207,7 +207,7 @@ impl Engine {
 			}
 		}
 
-		let inspect_path = format!("/v4.0.0/libpod/exec/{}/json", urlencoded(&exec_id));
+		let inspect_path = format!("{API_PREFIX}/exec/{}/json", urlencoded(&exec_id));
 		let inspect: ExecInspect = self
 			.client
 			.get_json(&inspect_path)
@@ -243,7 +243,7 @@ impl Engine {
 		let label = format!("podup.project={}", self.project);
 		let filters = serde_json::json!({ "label": [label] });
 		let path = format!(
-			"/v4.0.0/libpod/containers/json?all=true&filters={}",
+			"{API_PREFIX}/containers/json?all=true&filters={}",
 			urlencoded(&filters.to_string()),
 		);
 
@@ -265,7 +265,7 @@ impl Engine {
 				if !known.contains(name) {
 					tracing::info!("removing orphan container {name}");
 					let rm_path =
-						format!("/v4.0.0/libpod/containers/{}?force=true", urlencoded(name));
+						format!("{API_PREFIX}/containers/{}?force=true", urlencoded(name));
 					if let Err(e) = self.client.delete_ok(&rm_path).await {
 						tracing::debug!("orphan delete {name}: {e}");
 					}
@@ -294,7 +294,7 @@ impl Engine {
 			let service = &file.services[name];
 			for container_name in self.replica_names(name, service) {
 				let path = format!(
-					"/v4.0.0/libpod/containers/{}/top",
+					"{API_PREFIX}/containers/{}/top",
 					urlencoded(&container_name),
 				);
 				match self
@@ -337,7 +337,7 @@ impl Engine {
 		let container_name = self.first_replica_name(service_name, service);
 
 		let path = format!(
-			"/v4.0.0/libpod/containers/{}/json",
+			"{API_PREFIX}/containers/{}/json",
 			urlencoded(&container_name),
 		);
 		let info = self
@@ -375,7 +375,7 @@ impl Engine {
 				None if service.build.is_some() => format!("{name}:latest"),
 				None => continue,
 			};
-			let path = format!("/v4.0.0/libpod/images/{}/json", urlencoded(&image_ref));
+			let path = format!("{API_PREFIX}/images/{}/json", urlencoded(&image_ref));
 			match self.client.get_json::<ImageInspect>(&path).await {
 				Ok(img) => {
 					let (repo, tag) = image_ref
@@ -422,7 +422,7 @@ impl Engine {
 			.map(|(display, cname, is_tty)| {
 				let prefix = display.clone();
 				let path = format!(
-					"/v4.0.0/libpod/containers/{}/logs?stdout=true&stderr=true&follow=true",
+					"{API_PREFIX}/containers/{}/logs?stdout=true&stderr=true&follow=true",
 					urlencoded(cname),
 				);
 				let client = &self.client;
