+++ internal/lib.rs
@@ -15,6 +15,6 @@ pub mod size;
 pub mod substitute;
 
 pub use compose::{parse_file, parse_str, parse_str_raw, resolve_order};
-pub use engine::{Engine, RunOptions};
+pub use engine::{Engine, ProjectLock, RunOptions};
 pub use error::{ComposeError, Result};
 pub use libpod::Client;
