+++ internal/substitute/mod.rs
@@ -3,10 +3,14 @@
 //! Applies `${VAR}` / `$VAR` substitution to raw YAML text before parsing.
 //! Handles all compose-spec modifier forms: `:-`, `-`, `:+`, `+`, `:?`, `?`.
 
+mod parse;
+
 use std::collections::HashMap;
 use std::path::Path;
 
-use crate::error::{ComposeError, Result};
+use crate::error::Result;
+
+use parse::{collect_var_name, is_var_start, parse_braced_var, resolve_modifier};
 
 // ---------------------------------------------------------------------------
 // Public API
@@ -151,153 +155,6 @@ fn strip_dotenv_quotes(s: &str) -> &str {
 		return &s[1..s.len() - 1];
 	}
 	s
-}
-
-fn is_var_start(c: char) -> bool {
-	c.is_alphabetic() || c == '_'
-}
-
-fn is_var_char(c: char) -> bool {
-	c.is_alphanumeric() || c == '_'
-}
-
-fn collect_var_name(chars: &mut std::iter::Peekable<std::str::Chars<'_>>) -> String {
-	let mut name = String::new();
-	while let Some(&c) = chars.peek() {
-		if is_var_char(c) {
-			name.push(c);
-			chars.next();
-		} else {
-			break;
-		}
-	}
-	name
-}
-
-#[derive(Debug)]
-enum Modifier {
-	None,
-	/// `${VAR:-default}` — use default if unset or empty
-	DefaultIfUnsetOrEmpty(String),
-	/// `${VAR-default}` — use default if unset (empty value is OK)
-	DefaultIfUnset(String),
-	/// `${VAR:+value}` — use value if set and non-empty
-	AltIfSetAndNonEmpty(String),
-	/// `${VAR+value}` — use value if set (even if empty)
-	AltIfSet(String),
-	/// `${VAR:?error}` — error if unset or empty
-	ErrorIfUnsetOrEmpty(String),
-	/// `${VAR?error}` — error if unset
-	ErrorIfUnset(String),
-}
-
-/// Parse the content inside `${…}`.  The opening `{` has already been consumed.
-fn parse_braced_var(
-	chars: &mut std::iter::Peekable<std::str::Chars<'_>>,
-) -> Result<(String, Modifier)> {
-	let mut name = String::new();
-
-	loop {
-		match chars.peek() {
-			None => {
-				return Ok((name, Modifier::None));
-			}
-			Some('}') => {
-				chars.next();
-				return Ok((name, Modifier::None));
-			}
-			Some(':') => {
-				chars.next();
-				// Peek at what follows `:`.
-				let modifier = match chars.peek() {
-					Some('-') => {
-						chars.next();
-						Modifier::DefaultIfUnsetOrEmpty(collect_until_close(chars))
-					}
-					Some('+') => {
-						chars.next();
-						Modifier::AltIfSetAndNonEmpty(collect_until_close(chars))
-					}
-					Some('?') => {
-						chars.next();
-						Modifier::ErrorIfUnsetOrEmpty(collect_until_close(chars))
-					}
-					_ => Modifier::DefaultIfUnsetOrEmpty(collect_until_close(chars)),
-				};
-				return Ok((name, modifier));
-			}
-			Some('-') => {
-				chars.next();
-				return Ok((name, Modifier::DefaultIfUnset(collect_until_close(chars))));
-			}
-			Some('+') => {
-				chars.next();
-				return Ok((name, Modifier::AltIfSet(collect_until_close(chars))));
-			}
-			Some('?') => {
-				chars.next();
-				return Ok((name, Modifier::ErrorIfUnset(collect_until_close(chars))));
-			}
-			Some(&c) => {
-				name.push(c);
-				chars.next();
-			}
-		}
-	}
-}
-
-/// Collect everything until `}` (exclusive), consuming the `}`.
-fn collect_until_close(chars: &mut std::iter::Peekable<std::str::Chars<'_>>) -> String {
-	let mut buf = String::new();
-	for c in chars.by_ref() {
-		if c == '}' {
-			break;
-		}
-		buf.push(c);
-	}
-	buf
-}
-
-fn resolve_modifier(
-	var: String,
-	modifier: Modifier,
-	vars: &HashMap<String, String>,
-) -> Result<String> {
-	let value = vars.get(&var);
-
-	match modifier {
-		Modifier::None => Ok(value.cloned().unwrap_or_default()),
-
-		Modifier::DefaultIfUnsetOrEmpty(default) => match value {
-			Some(v) if !v.is_empty() => Ok(v.clone()),
-			_ => Ok(default),
-		},
-
-		Modifier::DefaultIfUnset(default) => match value {
-			Some(v) => Ok(v.clone()),
-			None => Ok(default),
-		},
-
-		Modifier::AltIfSetAndNonEmpty(alt) => match value {
-			Some(v) if !v.is_empty() => Ok(alt),
-			_ => Ok(String::new()),
-		},
-
-		Modifier::AltIfSet(alt) => match value {
-			Some(_) => Ok(alt),
-			None => Ok(String::new()),
-		},
-
-		Modifier::ErrorIfUnsetOrEmpty(msg) => match value {
-			Some(v) if !v.is_empty() => Ok(v.clone()),
-			_ => Err(ComposeError::RequiredVarNotSet { var, msg }),
-		},
-
-		Modifier::ErrorIfUnset(msg) => match value {
-			Some(v) => Ok(v.clone()),
-			None => Err(ComposeError::RequiredVarNotSet { var, msg }),
-		},
-	}
 }
 
 // ---------------------------------------------------------------------------
