+++ tests/engine_integration.rs
@@ -1,8 +1,12 @@
-/// Integration tests that exercise the engine against a real Podman daemon.
-///
-/// All tests skip gracefully when Podman is not reachable. In CI the
-/// `podman` input to the rust-ci reusable starts the socket and sets
-/// `PODMAN_SOCKET` before the coverage gate runs.
+//! Integration tests that exercise the engine against a real Podman daemon.
+//!
+//! All tests skip gracefully when Podman is not reachable. In CI the
+//! `podman` input to the rust-ci reusable starts the socket and sets
+//! `PODMAN_SOCKET` before the coverage gate runs.
+//!
+//! The test bodies are split across the `engine_integration/` submodules to
+//! keep each file under the source line limit. Shared helpers live here at the
+//! crate root so the submodules can reach them via `use super::*;`.
 use std::fs;
 
 use podup::{parse_str, Client, Engine};
@@ -24,2618 +28,31 @@ fn proj(tag: &str) -> String {
 	format!("t{}-{}", std::process::id(), tag)
 }
 
-// ---------------------------------------------------------------------------
-// Lifecycle
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn up_and_down() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("udn");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn up_no_recreate_skips_running() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("nor");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	// Second up with no_recreate: already running → skip
-	engine
-		.up_with_options(&file, false, &[], &[], true)
-		.await
-		.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn up_target_services_only() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("tgt");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  db:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    depends_on:\n      - db\n",
-	)
-	.unwrap();
-
-	// Only start web (and its dep db)
-	engine
-		.up_with_options(&file, false, &[], &["web".to_string()], false)
-		.await
-		.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn down_with_remove_volumes() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("dvol");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(&format!(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    volumes:\n      - {proj}-data:/data\nvolumes:\n  {proj}-data:\n"
-	))
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down_with_options(&file, true).await.unwrap();
-}
-
-#[tokio::test]
-async fn restart_all_services() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("rsa");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.restart(&file, None).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn restart_specific_service() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("rss");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.restart(&file, Some("web")).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn restart_cascade_dep() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("rcd");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  db:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    depends_on:\n      db:\n        condition: service_started\n        restart: true\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	// Restarting db triggers cascade restart of web
-	engine.restart(&file, Some("db")).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn restart_unknown_service_fails() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("ruf");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-
-	let err = engine
-		.restart(&file, Some("nonexistent"))
-		.await
-		.unwrap_err();
-	assert!(matches!(err, podup::ComposeError::ServiceNotFound(_)));
+/// Path to the built `podup` binary, for the CLI tests.
+fn bin() -> &'static str {
+	env!("CARGO_BIN_EXE_podup")
 }
 
 // ---------------------------------------------------------------------------
-// Query
+// Test groups (see engine_integration/*.rs)
 // ---------------------------------------------------------------------------
 
-#[tokio::test]
-async fn ps_shows_running_container() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("ps");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.ps(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn logs_from_named_service() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("lgs");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sh\", \"-c\", \"echo hello && sleep infinity\"]\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.logs(&file, Some("web"), false).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn logs_all_services() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("lga");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.logs(&file, None, false).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn logs_unknown_service_fails() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("lgf");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-
-	let err = engine
-		.logs(&file, Some("nonexistent"), false)
-		.await
-		.unwrap_err();
-	assert!(matches!(err, podup::ComposeError::ServiceNotFound(_)));
-}
-
-#[tokio::test]
-async fn exec_command_in_container() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("exc");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine
-		.exec(&file, "web", vec!["echo".to_string(), "test".to_string()])
-		.await
-		.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn exec_unknown_service_fails() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("exf");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-
-	let err = engine
-		.exec(&file, "nonexistent", vec!["echo".to_string()])
-		.await
-		.unwrap_err();
-	assert!(matches!(err, podup::ComposeError::ServiceNotFound(_)));
-}
-
-#[tokio::test]
-async fn pull_images() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("pll");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-
-	engine.pull(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn remove_orphans_no_orphans() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("orp");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.remove_orphans(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn attach_logs_empty_attach_returns() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("atl");
-	let engine = Engine::new(client, proj.clone());
-	// attach: false — attach_logs finds no targets and returns immediately
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    attach: false\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.attach_logs(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Volumes, secrets, configs
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn named_volume_created_on_up() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("nvol");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(&format!(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    volumes:\n      - {proj}-data:/data\nvolumes:\n  {proj}-data:\n"
-	))
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down_with_options(&file, true).await.unwrap();
-}
-
-#[tokio::test]
-async fn inline_secret_materialized() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("sec");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    secrets:\n      - mysecret\nsecrets:\n  mysecret:\n    content: \"supersecret\"\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	// Verify secret was bind-mounted by exec-ing a read
-	engine
-		.exec(
-			&file,
-			"web",
-			vec!["cat".to_string(), "/run/secrets/mysecret".to_string()],
-		)
-		.await
-		.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn file_secret_bound() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let dir = tempfile::tempdir().unwrap();
-	let secret_file = dir.path().join("my_secret.txt");
-	fs::write(&secret_file, b"file-secret-content").unwrap();
-
-	let proj = proj("fsec");
-	let engine = Engine::with_base_dir(client, proj.clone(), dir.path().to_path_buf());
-	let yaml = format!(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    secrets:\n      - filesecret\nsecrets:\n  filesecret:\n    file: {}\n",
-		secret_file.display()
-	);
-	let file = parse_str(&yaml).unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[test]
-fn env_secret_materialized() {
-	let rt = tokio::runtime::Runtime::new().unwrap();
-	temp_env::with_var("PODUP_TEST_SECRET_VAR", Some("env-secret-value"), || {
-		rt.block_on(async {
-			let client = match podman().await {
-				Some(d) => d,
-				None => return,
-			};
-			let proj = proj("esec");
-			let engine = Engine::new(client, proj.clone());
-			let file = parse_str(
-				"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    secrets:\n      - envsecret\nsecrets:\n  envsecret:\n    environment: PODUP_TEST_SECRET_VAR\n",
-			)
-			.unwrap();
-
-			engine.up(&file).await.unwrap();
-			engine.down(&file).await.unwrap();
-		});
-	});
-}
-
-#[tokio::test]
-async fn external_secret_skipped() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("xsec");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    secrets:\n      - extsecret\nsecrets:\n  extsecret:\n    external: true\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn invalid_secret_name_rejected() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("isec");
-	let engine = Engine::new(client, proj.clone());
-	// Secret name with path traversal
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    secrets:\n      - evils\nsecrets:\n  evils:\n    content: bad\n",
-	);
-	// Parse succeeds; but engine should reject the traversal during up()
-	// We can't actually test a ".." name since parse_str would accept it.
-	// Instead, verify that a normal name works (already covered by inline_secret test).
-	// This test exercises the path-validation code via a valid name edge case.
-	if let Ok(f) = file {
-		let _ = engine.up(&f).await;
-		let _ = engine.down(&f).await;
-	}
-}
-
-#[tokio::test]
-async fn inline_config_materialized() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("cfg");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    configs:\n      - mycfg\nconfigs:\n  mycfg:\n    content: \"key=value\"\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Lifecycle hooks
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn post_start_and_pre_stop_hooks_run() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("hks");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    post_start:\n      - command: [\"echo\", \"started\"]\n    pre_stop:\n      - command: [\"echo\", \"stopping\"]\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Health / depends_on
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn depends_on_service_healthy() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("hlt");
-	let engine = Engine::new(client, proj.clone());
-	// db has a healthcheck (CMD true), web waits for it to be healthy
-	let file = parse_str(
-		"services:\n  db:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    healthcheck:\n      test: [\"CMD\", \"true\"]\n      interval: 1s\n      timeout: 1s\n      retries: 5\n      start_period: 0s\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    depends_on:\n      db:\n        condition: service_healthy\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn depends_on_service_completed() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("cmp");
-	let engine = Engine::new(client, proj.clone());
-	// init exits 0 quickly; app waits for it to complete
-	let file = parse_str(
-		"services:\n  init:\n    image: alpine:latest\n    command: [\"sh\", \"-c\", \"exit 0\"]\n  app:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    depends_on:\n      init:\n        condition: service_completed_successfully\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Profiles
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn profile_filtered_service_skipped() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("prf");
-	let engine = Engine::new(client, proj.clone());
-	// "debug" service has profile "debug" — not in active profiles → skipped
-	// "web" has no profiles → always runs
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n  debug:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    profiles: [\"debug\"]\n",
-	)
-	.unwrap();
-
-	engine
-		.up_with_options(&file, false, &[], &[], false)
-		.await
-		.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Replicas
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn scale_creates_multiple_replicas() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("rep");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  worker:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    deploy:\n      replicas: 2\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Depends-on: service_healthy with no healthcheck
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn depends_on_healthy_no_healthcheck_skips_wait() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("hns");
-	let engine = Engine::new(client, proj.clone());
-	// backend has no healthcheck; frontend depends on it with service_healthy.
-	// podup logs a debug message and skips the wait.
-	let file = parse_str(
-		"services:\n  backend:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n  frontend:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    depends_on:\n      backend:\n        condition: service_healthy\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// PS with port bindings
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn ps_with_port_bindings() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("psb");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    ports:\n      - \"19100:80\"\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.ps(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Query: attach_logs streaming and logs stderr
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn attach_logs_streams_container_output() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("als");
-	let engine = Engine::new(client, proj.clone());
-	// Container writes to stdout and stderr then exits; attach_logs should
-	// stream the output and return when the stream ends (join_all completes).
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sh\", \"-c\", \"echo out-line; echo err-line >&2\"]\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.attach_logs(&file).await.unwrap();
-	let _ = engine.down(&file).await;
-}
-
-#[tokio::test]
-async fn logs_with_stderr_output() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("lge");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sh\", \"-c\", \"echo error-msg >&2; sleep infinity\"]\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.logs(&file, Some("web"), false).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Configs: file and environment sources
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn file_config_bound() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let dir = tempfile::tempdir().unwrap();
-	let cfg_file = dir.path().join("app.conf");
-	fs::write(&cfg_file, b"key=from-file").unwrap();
-
-	let proj = proj("fcfg");
-	let engine = Engine::with_base_dir(client, proj.clone(), dir.path().to_path_buf());
-	let yaml = format!(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    configs:\n      - filecfg\nconfigs:\n  filecfg:\n    file: {}\n",
-		cfg_file.display()
-	);
-	let file = parse_str(&yaml).unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[test]
-fn env_config_materialized() {
-	let rt = tokio::runtime::Runtime::new().unwrap();
-	temp_env::with_var("PODUP_TEST_CFG_VAR", Some("cfg-from-env"), || {
-		rt.block_on(async {
-			let client = match podman().await {
-				Some(d) => d,
-				None => return,
-			};
-			let proj = proj("ecfg");
-			let engine = Engine::new(client, proj.clone());
-			let file = parse_str(
-				"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    configs:\n      - envcfg\nconfigs:\n  envcfg:\n    environment: PODUP_TEST_CFG_VAR\n",
-			)
-			.unwrap();
-
-			engine.up(&file).await.unwrap();
-			engine.down(&file).await.unwrap();
-		});
-	});
-}
-
-// ---------------------------------------------------------------------------
-// Container options: expose, deploy labels, annotations, tmpfs long-form
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn service_with_expose_deploy_labels_annotations_tmpfs() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("sdl");
-	let engine = Engine::new(client, proj.clone());
-	// expose covers container.rs L56-63
-	// deploy.labels covers container.rs L76-78
-	// annotations covers container.rs L81-82
-	// long-form tmpfs volume covers container.rs L107-113 and L139
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    expose:\n      - \"8080\"\n    annotations:\n      com.example.note: value\n    deploy:\n      labels:\n        com.example.env: test\n    volumes:\n      - type: tmpfs\n        target: /tmp/cache\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Volume: named volume with driver_opts
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn named_volume_with_driver_opts() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let dir = tempfile::tempdir().unwrap();
-	let proj = proj("vdo");
-	let engine = Engine::with_base_dir(client, proj.clone(), dir.path().to_path_buf());
-	// driver_opts covers volume.rs L55 (Some(driver_opts) branch)
-	// Use a bind-mount volume pointing to the temp dir (fast, rootless-safe)
-	let yaml = format!(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    volumes:\n      - {proj}-cache:/cache\nvolumes:\n  {proj}-cache:\n    driver: local\n    driver_opts:\n      type: none\n      o: bind\n      device: {path}\n",
-		path = dir.path().display()
-	);
-	let file = parse_str(&yaml).unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down_with_options(&file, true).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Build
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn build_with_target_stage() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let dir = tempfile::tempdir().unwrap();
-	// Multi-stage Dockerfile — build with target: base covers build.rs L77
-	fs::write(
-		dir.path().join("Dockerfile"),
-		b"FROM alpine:latest AS base\nRUN echo base-stage\nFROM base AS final\nRUN echo final-stage\n",
-	)
-	.unwrap();
-
-	let proj = proj("bst");
-	let engine = Engine::with_base_dir(client, proj.clone(), dir.path().to_path_buf());
-	let image_tag = format!("podup-test-bst-{}:latest", std::process::id());
-	let yaml = format!(
-		"services:\n  app:\n    build:\n      context: .\n      target: base\n    image: {image_tag}\n    command: [\"sleep\", \"infinity\"]\n"
-	);
-	let file = parse_str(&yaml).unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn build_with_args_and_extra_tags() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let dir = tempfile::tempdir().unwrap();
-	let proj = proj("bat");
-	let engine = Engine::with_base_dir(client, proj.clone(), dir.path().to_path_buf());
-	let pid = std::process::id();
-	let main_tag = format!("podup-test-bat-{}:latest", pid);
-	let extra_tag = format!("podup-test-bat-extra-{}:v1", pid);
-	let yaml = format!(
-		"services:\n  app:\n    build:\n      context: .\n      dockerfile_inline: |\n        FROM alpine:latest\n        ARG VERSION=0\n        RUN echo Version $VERSION\n      args:\n        VERSION: \"1.0\"\n      tags:\n        - {extra_tag}\n    image: {main_tag}\n    command: [\"sleep\", \"infinity\"]\n"
-	);
-	let file = parse_str(&yaml).unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn build_inline_dockerfile() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let dir = tempfile::tempdir().unwrap();
-	let proj = proj("bld");
-	let engine = Engine::with_base_dir(client, proj.clone(), dir.path().to_path_buf());
-	let image_tag = format!("podup-test-build-{}:latest", std::process::id());
-	let yaml = format!(
-		"services:\n  app:\n    build:\n      context: .\n      dockerfile_inline: |\n        FROM alpine:latest\n        RUN echo built\n    image: {image_tag}\n    command: [\"sleep\", \"infinity\"]\n"
-	);
-	let file = parse_str(&yaml).unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn build_from_dockerfile_in_context() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let dir = tempfile::tempdir().unwrap();
-	fs::write(
-		dir.path().join("Dockerfile"),
-		b"FROM alpine:latest\nRUN echo context-build\n",
-	)
-	.unwrap();
-
-	let proj = proj("bdc");
-	let engine = Engine::with_base_dir(client, proj.clone(), dir.path().to_path_buf());
-	let image_tag = format!("podup-test-build-ctx-{}:latest", std::process::id());
-	let yaml = format!(
-		"services:\n  app:\n    build:\n      context: .\n    image: {image_tag}\n    command: [\"sleep\", \"infinity\"]\n"
-	);
-	let file = parse_str(&yaml).unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Networks
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn explicit_network_created() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("net");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    networks:\n      - mynet\nnetworks:\n  mynet:\n    driver: bridge\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Secret/config long-form refs
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn secret_long_form_ref() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("slf");
-	let engine = Engine::new(client, proj.clone());
-	// mode: 256 = 0o400; uid exercises apply_owner (best-effort chown)
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    secrets:\n      - source: mysecret\n        target: /run/secrets/custom_name\n        mode: 256\n        uid: \"0\"\nsecrets:\n  mysecret:\n    content: \"topsecret\"\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine
-		.exec(
-			&file,
-			"web",
-			vec!["cat".to_string(), "/run/secrets/custom_name".to_string()],
-		)
-		.await
-		.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn config_long_form_ref() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("clf");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    configs:\n      - source: mycfg\n        target: /etc/app.conf\nconfigs:\n  mycfg:\n    content: \"key=value\"\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine
-		.exec(
-			&file,
-			"web",
-			vec!["cat".to_string(), "/etc/app.conf".to_string()],
-		)
-		.await
-		.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// External volume skip
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn external_volume_skipped_on_up() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("exv");
-	let engine = Engine::new(client, proj.clone());
-	// The external volume is declared but not mounted by the service,
-	// so create_volumes() hits the `continue` branch without creating it.
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\nvolumes:\n  extdata:\n    external: true\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Orphan removal
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn remove_orphans_removes_container() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("orr");
-	let engine = Engine::new(client, proj.clone());
-
-	let file_svc1 = parse_str(
-		"services:\n  svc1:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-	engine.up(&file_svc1).await.unwrap();
-
-	// file_svc2 only declares svc2 — svc1 becomes an orphan
-	let file_svc2 = parse_str(
-		"services:\n  svc2:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-	engine.remove_orphans(&file_svc2).await.unwrap();
-
-	// cleanup (svc1 already removed; down() on either file is a no-op for missing containers)
-	let _ = engine.down(&file_svc1).await;
-}
-
-// ---------------------------------------------------------------------------
-// Health: non-zero exit
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn wait_completed_nonzero_error() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("cne");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  init:\n    image: alpine:latest\n    command: [\"sh\", \"-c\", \"exit 1\"]\n  app:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    depends_on:\n      init:\n        condition: service_completed_successfully\n",
-	)
-	.unwrap();
-
-	// up() propagates the non-zero exit error from wait_completed
-	let err = engine.up(&file).await.unwrap_err();
-	assert!(
-		matches!(err, podup::ComposeError::HealthCheckTimeout(_)),
-		"expected HealthCheckTimeout, got: {err}"
-	);
-	let _ = engine.down(&file).await;
-}
-
-// ---------------------------------------------------------------------------
-// Profiles: COMPOSE_PROFILES env var path
-// ---------------------------------------------------------------------------
-
-#[test]
-fn active_profiles_via_env() {
-	let rt = tokio::runtime::Runtime::new().unwrap();
-	// Set COMPOSE_PROFILES so active_profiles_set reads it (covers profiles.rs L15-19)
-	temp_env::with_var("COMPOSE_PROFILES", Some("prod"), || {
-		rt.block_on(async {
-			let client = match podman().await {
-				Some(d) => d,
-				None => return,
-			};
-			let proj = proj("apv");
-			let engine = Engine::new(client, proj.clone());
-			// "debug" service has profile "debug" — not in "prod" → skipped
-			let file = parse_str(
-				"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n  debug:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    profiles: [\"debug\"]\n",
-			)
-			.unwrap();
-			// Pass empty active_profiles slice so it falls back to COMPOSE_PROFILES env
-			engine
-				.up_with_options(&file, false, &[], &[], false)
-				.await
-				.unwrap();
-			engine.down(&file).await.unwrap();
-		});
-	});
-}
-
-// ---------------------------------------------------------------------------
-// Health: wait_healthy timeout and wait_completed polling
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn wait_healthy_times_out() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("wht");
-	let engine = Engine::new(client, proj.clone());
-	// CMD false always fails; retries:1 means wait_healthy exhausts quickly
-	// Covers health.rs L42-43 (loop body closing braces) and L47 (timeout Err)
-	let file = parse_str(
-		"services:\n  db:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    healthcheck:\n      test: [\"CMD\", \"false\"]\n      interval: 1s\n      timeout: 1s\n      retries: 1\n      start_period: 0s\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    depends_on:\n      db:\n        condition: service_healthy\n",
-	)
-	.unwrap();
-
-	let err = engine.up(&file).await.unwrap_err();
-	assert!(
-		matches!(err, podup::ComposeError::HealthCheckTimeout(_)),
-		"expected HealthCheckTimeout, got: {err}"
-	);
-	let _ = engine.down(&file).await;
-}
-
-#[tokio::test]
-async fn wait_completed_polling() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("wcp");
-	let engine = Engine::new(client, proj.clone());
-	// init sleeps 1.5s before exiting; first poll sees it running (L73-75 covered)
-	let file = parse_str(
-		"services:\n  init:\n    image: alpine:latest\n    command: [\"sh\", \"-c\", \"sleep 1.5; exit 0\"]\n  app:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    depends_on:\n      init:\n        condition: service_completed_successfully\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// External config skipped
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn external_config_skipped() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("xcfg");
-	let engine = Engine::new(client, proj.clone());
-	// Covers volume.rs L215 (external config debug log)
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    configs:\n      - extcfg\nconfigs:\n  extcfg:\n    external: true\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Container options: expose with slash, env_file, ulimits
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn service_with_expose_proto_and_ulimits() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	// expose "8080/tcp" (with slash) covers container.rs L57 (raw.clone() branch)
-	// ulimits covers container.rs L150 (Some(ulimits))
-	let proj = proj("seu");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    expose:\n      - \"8080/tcp\"\n    ulimits:\n      nofile:\n        soft: 1024\n        hard: 2048\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn env_file_loaded() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let dir = tempfile::tempdir().unwrap();
-	fs::write(dir.path().join("test.env"), b"MYVAR=hello\n").unwrap();
-
-	let proj = proj("evf");
-	let engine = Engine::with_base_dir(client, proj.clone(), dir.path().to_path_buf());
-	// env_file covers container.rs L278 (load_env_file_entries)
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    env_file:\n      - test.env\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Lifecycle: target_set skips non-targeted service; dep profile skip
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn target_services_skips_non_dep() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("tsk");
-	let engine = Engine::new(client, proj.clone());
-	// "extra" is not depended upon by web → skipped (lifecycle.rs L56 continue)
-	let file = parse_str(
-		"services:\n  extra:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-
-	engine
-		.up_with_options(&file, false, &[], &["web".to_string()], false)
-		.await
-		.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn dep_on_profile_filtered_service() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("dpf");
-	let engine = Engine::new(client, proj.clone());
-	// "db" has profile "debug" → not active → dep wait skipped (lifecycle.rs L73)
-	// "web" depends on "db" but db is profile-filtered so its dep wait is skipped
-	let file = parse_str(
-		"services:\n  db:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    profiles: [\"debug\"]\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    depends_on:\n      - db\n",
-	)
-	.unwrap();
-
-	// No active profiles → db is skipped; web still runs but skips db's dep wait
-	engine
-		.up_with_options(&file, false, &[], &[], false)
-		.await
-		.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Build: arg with null value (from environment)
-// ---------------------------------------------------------------------------
-
-#[test]
-fn build_with_env_arg() {
-	let rt = tokio::runtime::Runtime::new().unwrap();
-	// FROM_ENV has no explicit value → read from environment (build.rs L89 None branch)
-	temp_env::with_var("FROM_ENV", Some("test-value"), || {
-		rt.block_on(async {
-			let client = match podman().await {
-				Some(d) => d,
-				None => return,
-			};
-			let dir = tempfile::tempdir().unwrap();
-			let proj = proj("bea");
-			let engine = Engine::with_base_dir(client, proj.clone(), dir.path().to_path_buf());
-			let image_tag = format!("podup-test-bea-{}:latest", std::process::id());
-			let yaml = format!(
-				"services:\n  app:\n    build:\n      context: .\n      dockerfile_inline: |\n        FROM alpine:latest\n        ARG FROM_ENV\n        RUN echo env=$FROM_ENV\n      args:\n        FROM_ENV:\n    image: {image_tag}\n    command: [\"sleep\", \"infinity\"]\n"
-			);
-			let file = parse_str(&yaml).unwrap();
-
-			engine.up(&file).await.unwrap();
-			engine.down(&file).await.unwrap();
-		});
-	});
-}
-
-// ---------------------------------------------------------------------------
-// label_file: load labels from file (container.rs L73-74)
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn label_file_labels_applied() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let dir = tempfile::tempdir().unwrap();
-	fs::write(
-		dir.path().join("svc.labels"),
-		b"com.example.role=web\ncom.example.env=test\n",
-	)
-	.unwrap();
-	let proj = proj("lfl");
-	let engine = Engine::with_base_dir(client, proj.clone(), dir.path().to_path_buf());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    label_file: svc.labels\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// optional dep not in file (lifecycle.rs L45, L70)
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn optional_dep_not_in_file() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("odf");
-	let engine = Engine::new(client, proj.clone());
-	// ghost_db not in services + required:false → resolve_order skips it,
-	// target_set pushes it (file.services.get → None → L45),
-	// dep-wait loop hits None => continue (L70)
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    depends_on:\n      ghost_db:\n        condition: service_started\n        required: false\n",
-	)
-	.unwrap();
-
-	engine
-		.up_with_options(&file, false, &[], &["web".to_string()], false)
-		.await
-		.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// duplicate target_services triggers continue in target_set (lifecycle.rs L37)
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn target_services_duplicate_entry() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("tde");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-
-	// Passing "web" twice causes it to be pushed to the target_set stack twice;
-	// the second pop finds "web" already in the set → !set.insert → continue (L37).
-	engine
-		.up_with_options(
-			&file,
-			false,
-			&[],
-			&["web".to_string(), "web".to_string()],
-			false,
-		)
-		.await
-		.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Watch (requires test-helpers feature)
-// ---------------------------------------------------------------------------
+#[path = "engine_integration/group_a1.rs"]
+mod group_a1;
+#[path = "engine_integration/group_a2.rs"]
+mod group_a2;
+#[path = "engine_integration/group_a3.rs"]
+mod group_a3;
+#[path = "engine_integration/group_a4.rs"]
+mod group_a4;
+#[path = "engine_integration/group_b1.rs"]
+mod group_b1;
 
 #[cfg(feature = "test-helpers")]
-mod watch_tests {
-	use std::time::Duration;
-
-	use super::*;
-
-	#[tokio::test]
-	async fn watch_no_develop_rules_returns_immediately() {
-		let client = match podman().await {
-			Some(d) => d,
-			None => return,
-		};
-		let proj = proj("wno");
-		let engine = Engine::new(client, proj.clone());
-		// No develop.watch section → watch() returns Ok(()) immediately
-		let file = parse_str(
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-		)
-		.unwrap();
-
-		engine.watch(&file).await.unwrap();
-	}
-
-	#[tokio::test]
-	async fn watch_sync_file_to_container() {
-		let client = match podman().await {
-			Some(d) => d,
-			None => return,
-		};
-		let dir = tempfile::tempdir().unwrap();
-		let src_file = dir.path().join("app.txt");
-		fs::write(&src_file, b"initial content").unwrap();
-
-		let proj = proj("wsy");
-		let engine = Engine::with_base_dir(client, proj.clone(), dir.path().to_path_buf());
-		let file = parse_str(
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-		)
-		.unwrap();
-
-		engine.up(&file).await.unwrap();
-		engine
-			.test_sync_to_container(&format!("{proj}-web"), &src_file, "/tmp/app.txt")
-			.await
-			.unwrap();
-		engine.down(&file).await.unwrap();
-	}
-
-	#[tokio::test]
-	async fn watch_restart_container() {
-		let client = match podman().await {
-			Some(d) => d,
-			None => return,
-		};
-		let proj = proj("wrs");
-		let engine = Engine::new(client, proj.clone());
-		let file = parse_str(
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-		)
-		.unwrap();
-
-		engine.up(&file).await.unwrap();
-		engine
-			.test_watch_restart(&format!("{proj}-web"))
-			.await
-			.unwrap();
-		engine.down(&file).await.unwrap();
-	}
-
-	#[tokio::test]
-	async fn watch_exec_in_container() {
-		let client = match podman().await {
-			Some(d) => d,
-			None => return,
-		};
-		let proj = proj("wex");
-		let engine = Engine::new(client, proj.clone());
-		let file = parse_str(
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-		)
-		.unwrap();
-
-		engine.up(&file).await.unwrap();
-		engine
-			.test_watch_exec(
-				&format!("{proj}-web"),
-				vec!["echo".to_string(), "from-watch-exec".to_string()],
-			)
-			.await
-			.unwrap();
-		engine.down(&file).await.unwrap();
-	}
-
-	#[tokio::test]
-	async fn watch_initial_sync_runs() {
-		let client = match podman().await {
-			Some(d) => d,
-			None => return,
-		};
-		let dir = tempfile::tempdir().unwrap();
-		let src = dir.path().join("app.txt");
-		fs::write(&src, b"initial").unwrap();
-
-		let proj = proj("wis");
-		let engine = Engine::with_base_dir(client, proj.clone(), dir.path().to_path_buf());
-		let file = parse_str(
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    develop:\n      watch:\n        - path: app.txt\n          action: sync\n          target: /tmp/app.txt\n          initial_sync: true\n",
-		)
-		.unwrap();
-
-		engine.up(&file).await.unwrap();
-
-		let client2 = podup::podman::connect_from_env()
-			.or_else(|_| podup::podman::connect(None))
-			.unwrap();
-		let engine2 = Engine::with_base_dir(client2, proj.clone(), dir.path().to_path_buf());
-		let file2 = file.clone();
-		let handle = tokio::spawn(async move { engine2.watch(&file2).await });
-		// Give watch() time to run initial_sync before aborting
-		tokio::time::sleep(Duration::from_millis(300)).await;
-		handle.abort();
-
-		engine.down(&file).await.unwrap();
-	}
-
-	#[tokio::test]
-	async fn watch_restart_action_via_event() {
-		let client = match podman().await {
-			Some(d) => d,
-			None => return,
-		};
-		let dir = tempfile::tempdir().unwrap();
-		let watch_dir = dir.path().join("src");
-		fs::create_dir(&watch_dir).unwrap();
-		fs::write(watch_dir.join("main.txt"), b"v1").unwrap();
-
-		let proj = proj("wra");
-		let engine = Engine::with_base_dir(client, proj.clone(), dir.path().to_path_buf());
-		let file = parse_str(
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    develop:\n      watch:\n        - path: src\n          action: restart\n",
-		)
-		.unwrap();
-
-		engine.up(&file).await.unwrap();
-
-		let client2 = podup::podman::connect_from_env()
-			.or_else(|_| podup::podman::connect(None))
-			.unwrap();
-		let engine2 = Engine::with_base_dir(client2, proj.clone(), dir.path().to_path_buf());
-		let file2 = file.clone();
-		let handle = tokio::spawn(async move { engine2.watch(&file2).await });
-
-		tokio::time::sleep(Duration::from_millis(150)).await;
-		fs::write(watch_dir.join("main.txt"), b"v2").unwrap();
-		tokio::time::sleep(Duration::from_millis(400)).await;
-
-		handle.abort();
-		engine.down(&file).await.unwrap();
-	}
-
-	#[tokio::test]
-	async fn watch_sync_and_restart_action_via_event() {
-		let client = match podman().await {
-			Some(d) => d,
-			None => return,
-		};
-		let dir = tempfile::tempdir().unwrap();
-		let watch_dir = dir.path().join("src");
-		fs::create_dir(&watch_dir).unwrap();
-		fs::write(watch_dir.join("main.txt"), b"v1").unwrap();
-
-		let proj = proj("wsr");
-		let engine = Engine::with_base_dir(client, proj.clone(), dir.path().to_path_buf());
-		let file = parse_str(
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    develop:\n      watch:\n        - path: src\n          action: sync+restart\n          target: /app/\n",
-		)
-		.unwrap();
-
-		engine.up(&file).await.unwrap();
-
-		let client2 = podup::podman::connect_from_env()
-			.or_else(|_| podup::podman::connect(None))
-			.unwrap();
-		let engine2 = Engine::with_base_dir(client2, proj.clone(), dir.path().to_path_buf());
-		let file2 = file.clone();
-		let handle = tokio::spawn(async move { engine2.watch(&file2).await });
-
-		tokio::time::sleep(Duration::from_millis(150)).await;
-		fs::write(watch_dir.join("main.txt"), b"v2").unwrap();
-		tokio::time::sleep(Duration::from_millis(400)).await;
-
-		handle.abort();
-		engine.down(&file).await.unwrap();
-	}
-
-	#[tokio::test]
-	async fn watch_sync_and_exec_action_via_event() {
-		let client = match podman().await {
-			Some(d) => d,
-			None => return,
-		};
-		let dir = tempfile::tempdir().unwrap();
-		let watch_dir = dir.path().join("src");
-		fs::create_dir(&watch_dir).unwrap();
-		fs::write(watch_dir.join("main.txt"), b"v1").unwrap();
-
-		let proj = proj("wse");
-		let engine = Engine::with_base_dir(client, proj.clone(), dir.path().to_path_buf());
-		let file = parse_str(
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    develop:\n      watch:\n        - path: src\n          action: sync+exec\n          target: /app/\n          exec:\n            command: [\"echo\", \"reloaded\"]\n",
-		)
-		.unwrap();
-
-		engine.up(&file).await.unwrap();
-
-		let client2 = podup::podman::connect_from_env()
-			.or_else(|_| podup::podman::connect(None))
-			.unwrap();
-		let engine2 = Engine::with_base_dir(client2, proj.clone(), dir.path().to_path_buf());
-		let file2 = file.clone();
-		let handle = tokio::spawn(async move { engine2.watch(&file2).await });
-
-		tokio::time::sleep(Duration::from_millis(150)).await;
-		fs::write(watch_dir.join("main.txt"), b"v2").unwrap();
-		tokio::time::sleep(Duration::from_millis(400)).await;
-
-		handle.abort();
-		engine.down(&file).await.unwrap();
-	}
-
-	#[tokio::test]
-	async fn watch_event_loop_dispatches_sync() {
-		let client = match podman().await {
-			Some(d) => d,
-			None => return,
-		};
-		let dir = tempfile::tempdir().unwrap();
-		let watch_dir = dir.path().join("src");
-		fs::create_dir(&watch_dir).unwrap();
-		fs::write(watch_dir.join("app.txt"), b"v1").unwrap();
-
-		let proj = proj("wev");
-		let engine = Engine::with_base_dir(client, proj.clone(), dir.path().to_path_buf());
-		let rel_path = "src";
-		let yaml = format!(
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    develop:\n      watch:\n        - path: {rel_path}\n          action: sync\n          target: /app/\n"
-		);
-		let file = parse_str(&yaml).unwrap();
-
-		engine.up(&file).await.unwrap();
-
-		let client2 = podup::podman::connect_from_env()
-			.or_else(|_| podup::podman::connect(None))
-			.unwrap();
-		let engine2 = Engine::with_base_dir(client2, proj.clone(), dir.path().to_path_buf());
-		let file2 = file.clone();
-		let watch_handle = tokio::spawn(async move { engine2.watch(&file2).await });
-
-		// Modify a file to trigger the event dispatch path
-		tokio::time::sleep(Duration::from_millis(150)).await;
-		fs::write(watch_dir.join("app.txt"), b"v2").unwrap();
-		tokio::time::sleep(Duration::from_millis(400)).await;
-
-		watch_handle.abort();
-		engine.down(&file).await.unwrap();
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Pause / unpause
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn pause_and_unpause() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("pau");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.pause(&file, &[]).await.unwrap();
-	engine.unpause(&file, &[]).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Run
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn engine_run_command_succeeds() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("run");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str("services:\n  job:\n    image: alpine:latest\n").unwrap();
-
-	let result = engine
-		.run(
-			&file,
-			"job",
-			podup::RunOptions {
-				cmd: vec!["echo".to_string(), "hello".to_string()],
-				rm: true,
-				detach: false,
-				env_overrides: vec![],
-				name_override: None,
-			},
-		)
-		.await;
-	assert!(result.is_ok(), "run failed: {result:?}");
-}
-
-#[tokio::test]
-async fn engine_run_nonzero_exit_returns_run_exited() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("rxc");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str("services:\n  job:\n    image: alpine:latest\n").unwrap();
-
-	let result = engine
-		.run(
-			&file,
-			"job",
-			podup::RunOptions {
-				cmd: vec!["false".to_string()],
-				rm: true,
-				detach: false,
-				env_overrides: vec![],
-				name_override: None,
-			},
-		)
-		.await;
-	assert!(
-		matches!(result, Err(podup::ComposeError::RunExited(_))),
-		"expected RunExited, got {result:?}"
-	);
-}
-
-// ---------------------------------------------------------------------------
-// Top
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn engine_top_running_container() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("top");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.top(&file, &[]).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Images
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn engine_images_lists_service_images() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("img");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.images(&file).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Port
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn engine_port_returns_binding() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("prt");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    ports:\n      - \"127.0.0.1:18080:80\"\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.port(&file, "web", 80, "tcp").await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Cp
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn engine_cp_from_container_extracts_file() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let dir = tempfile::tempdir().unwrap();
-	let proj = proj("cpf");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-
-	let dst = dir.path().to_str().unwrap().to_string();
-	let src = "web:/etc/hostname".to_string();
-	let result = engine.cp(&file, &src, &dst).await;
-	engine.down(&file).await.unwrap();
-
-	result.unwrap();
-	assert!(dir.path().join("hostname").exists());
-}
-
-#[tokio::test]
-async fn engine_cp_to_container_uploads_file() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let dir = tempfile::tempdir().unwrap();
-	let local_file = dir.path().join("testfile.txt");
-	fs::write(&local_file, b"hello from host").unwrap();
-
-	let proj = proj("cpt");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-
-	let src = local_file.to_str().unwrap().to_string();
-	let dst = "web:/tmp".to_string();
-	let result = engine.cp(&file, &src, &dst).await;
-	engine.down(&file).await.unwrap();
-
-	result.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// Replicas: restart, logs, top, exec, port target correct containers
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn restart_scaled_service_all_replicas() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("rsr");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  worker:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    deploy:\n      replicas: 2\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	// Both replicas must be reachable for restart to succeed.
-	engine.restart(&file, Some("worker")).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn logs_scaled_service_all_replicas() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("lsr");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  worker:\n    image: alpine:latest\n    command: [\"sh\", \"-c\", \"echo hello && sleep infinity\"]\n    deploy:\n      replicas: 2\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	// logs for a named service with replicas: should stream from all without error.
-	engine.logs(&file, Some("worker"), false).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn top_scaled_service_all_replicas() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("tsr");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  worker:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    deploy:\n      replicas: 2\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.top(&file, &[]).await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn exec_scaled_service_targets_first_replica() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("esr");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  worker:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    deploy:\n      replicas: 2\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine
-		.exec(&file, "worker", vec!["echo".to_string(), "ok".to_string()])
-		.await
-		.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-#[tokio::test]
-async fn port_scaled_service_targets_first_replica() {
-	let client = match podman().await {
-		Some(d) => d,
-		None => return,
-	};
-	let proj = proj("psr");
-	let engine = Engine::new(client, proj.clone());
-	let file = parse_str(
-		"services:\n  worker:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    ports:\n      - \"80\"\n    deploy:\n      replicas: 2\n",
-	)
-	.unwrap();
-
-	engine.up(&file).await.unwrap();
-	engine.port(&file, "worker", 80, "tcp").await.unwrap();
-	engine.down(&file).await.unwrap();
-}
-
-// ---------------------------------------------------------------------------
-// CLI binary (covers main.rs)
-// ---------------------------------------------------------------------------
-
-mod cli_tests {
-	use std::fs;
-	use std::process::Command;
-	use tempfile::tempdir;
-
-	fn bin() -> &'static str {
-		env!("CARGO_BIN_EXE_podup")
-	}
-
-	#[test]
-	fn cli_config_no_podman() {
-		let dir = tempdir().unwrap();
-		let compose = dir.path().join("docker-compose.yml");
-		fs::write(&compose, "services:\n  web:\n    image: alpine:latest\n").unwrap();
-
-		let out = Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "config"])
-			.output()
-			.expect("podup binary not found");
-
-		assert!(
-			out.status.success(),
-			"config failed: {}",
-			String::from_utf8_lossy(&out.stderr)
-		);
-		let stdout = String::from_utf8_lossy(&out.stdout);
-		assert!(stdout.contains("alpine"));
-	}
-
-	#[tokio::test]
-	async fn cli_up_and_down_via_binary() {
-		if super::podman().await.is_none() {
-			return;
-		}
-		let dir = tempdir().unwrap();
-		let compose = dir.path().join("docker-compose.yml");
-		let pid = std::process::id();
-		let proj = format!("t{}-cli", pid);
-		fs::write(
-			&compose,
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-		)
-		.unwrap();
-
-		let up = Command::new(bin())
-			.args([
-				"-f",
-				compose.to_str().unwrap(),
-				"-p",
-				&proj,
-				"up",
-				"--detach",
-			])
-			.output()
-			.unwrap();
-		assert!(
-			up.status.success(),
-			"up failed: {}",
-			String::from_utf8_lossy(&up.stderr)
-		);
-
-		let down = Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "down"])
-			.output()
-			.unwrap();
-		assert!(
-			down.status.success(),
-			"down failed: {}",
-			String::from_utf8_lossy(&down.stderr)
-		);
-	}
-
-	#[tokio::test]
-	async fn cli_ps_subcommand() {
-		if super::podman().await.is_none() {
-			return;
-		}
-		let dir = tempdir().unwrap();
-		let compose = dir.path().join("docker-compose.yml");
-		let proj = format!("t{}-clps", std::process::id());
-		fs::write(
-			&compose,
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-		)
-		.unwrap();
-
-		Command::new(bin())
-			.args([
-				"-f",
-				compose.to_str().unwrap(),
-				"-p",
-				&proj,
-				"up",
-				"--detach",
-			])
-			.output()
-			.unwrap();
-
-		let ps = Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "ps"])
-			.output()
-			.unwrap();
-		assert!(ps.status.success());
-
-		Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "down"])
-			.output()
-			.unwrap();
-	}
-
-	#[tokio::test]
-	async fn cli_logs_subcommand() {
-		if super::podman().await.is_none() {
-			return;
-		}
-		let dir = tempdir().unwrap();
-		let compose = dir.path().join("docker-compose.yml");
-		let proj = format!("t{}-cllg", std::process::id());
-		fs::write(
-			&compose,
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-		)
-		.unwrap();
-
-		Command::new(bin())
-			.args([
-				"-f",
-				compose.to_str().unwrap(),
-				"-p",
-				&proj,
-				"up",
-				"--detach",
-			])
-			.output()
-			.unwrap();
-
-		let logs = Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "logs"])
-			.output()
-			.unwrap();
-		assert!(logs.status.success());
-
-		Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "down"])
-			.output()
-			.unwrap();
-	}
-
-	#[tokio::test]
-	async fn cli_exec_subcommand() {
-		if super::podman().await.is_none() {
-			return;
-		}
-		let dir = tempdir().unwrap();
-		let compose = dir.path().join("docker-compose.yml");
-		let proj = format!("t{}-clex", std::process::id());
-		fs::write(
-			&compose,
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-		)
-		.unwrap();
-
-		Command::new(bin())
-			.args([
-				"-f",
-				compose.to_str().unwrap(),
-				"-p",
-				&proj,
-				"up",
-				"--detach",
-			])
-			.output()
-			.unwrap();
-
-		let exec = Command::new(bin())
-			.args([
-				"-f",
-				compose.to_str().unwrap(),
-				"-p",
-				&proj,
-				"exec",
-				"web",
-				"echo",
-				"cli-exec",
-			])
-			.output()
-			.unwrap();
-		assert!(exec.status.success());
-
-		Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "down"])
-			.output()
-			.unwrap();
-	}
-
-	#[tokio::test]
-	async fn cli_restart_subcommand() {
-		if super::podman().await.is_none() {
-			return;
-		}
-		let dir = tempdir().unwrap();
-		let compose = dir.path().join("docker-compose.yml");
-		let proj = format!("t{}-clrs", std::process::id());
-		fs::write(
-			&compose,
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-		)
-		.unwrap();
-
-		Command::new(bin())
-			.args([
-				"-f",
-				compose.to_str().unwrap(),
-				"-p",
-				&proj,
-				"up",
-				"--detach",
-			])
-			.output()
-			.unwrap();
-
-		let restart = Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "restart"])
-			.output()
-			.unwrap();
-		assert!(restart.status.success());
-
-		Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "down"])
-			.output()
-			.unwrap();
-	}
-
-	#[tokio::test]
-	async fn cli_pull_subcommand() {
-		if super::podman().await.is_none() {
-			return;
-		}
-		let dir = tempdir().unwrap();
-		let compose = dir.path().join("docker-compose.yml");
-		fs::write(&compose, "services:\n  web:\n    image: alpine:latest\n").unwrap();
-
-		let pull = Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "pull"])
-			.output()
-			.unwrap();
-		assert!(pull.status.success());
-	}
-
-	#[tokio::test]
-	async fn cli_stop_and_start_subcommands() {
-		if super::podman().await.is_none() {
-			return;
-		}
-		let dir = tempdir().unwrap();
-		let compose = dir.path().join("docker-compose.yml");
-		let proj = format!("t{}-clss", std::process::id());
-		fs::write(
-			&compose,
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-		)
-		.unwrap();
-
-		Command::new(bin())
-			.args([
-				"-f",
-				compose.to_str().unwrap(),
-				"-p",
-				&proj,
-				"up",
-				"--detach",
-			])
-			.output()
-			.unwrap();
-
-		let stop = Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "stop"])
-			.output()
-			.unwrap();
-		assert!(stop.status.success(), "stop failed: {:?}", stop.stderr);
-
-		let start = Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "start"])
-			.output()
-			.unwrap();
-		assert!(start.status.success(), "start failed: {:?}", start.stderr);
-
-		Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "down"])
-			.output()
-			.unwrap();
-	}
-
-	#[tokio::test]
-	async fn cli_kill_subcommand() {
-		if super::podman().await.is_none() {
-			return;
-		}
-		let dir = tempdir().unwrap();
-		let compose = dir.path().join("docker-compose.yml");
-		let proj = format!("t{}-clkl", std::process::id());
-		fs::write(
-			&compose,
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-		)
-		.unwrap();
-
-		Command::new(bin())
-			.args([
-				"-f",
-				compose.to_str().unwrap(),
-				"-p",
-				&proj,
-				"up",
-				"--detach",
-			])
-			.output()
-			.unwrap();
-
-		let kill = Command::new(bin())
-			.args([
-				"-f",
-				compose.to_str().unwrap(),
-				"-p",
-				&proj,
-				"kill",
-				"--signal",
-				"SIGTERM",
-			])
-			.output()
-			.unwrap();
-		assert!(kill.status.success(), "kill failed: {:?}", kill.stderr);
-
-		Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "down"])
-			.output()
-			.unwrap();
-	}
-
-	#[tokio::test]
-	async fn cli_rm_subcommand() {
-		if super::podman().await.is_none() {
-			return;
-		}
-		let dir = tempdir().unwrap();
-		let compose = dir.path().join("docker-compose.yml");
-		let proj = format!("t{}-clrm", std::process::id());
-		fs::write(
-			&compose,
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-		)
-		.unwrap();
-
-		Command::new(bin())
-			.args([
-				"-f",
-				compose.to_str().unwrap(),
-				"-p",
-				&proj,
-				"up",
-				"--detach",
-			])
-			.output()
-			.unwrap();
-
-		Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "stop"])
-			.output()
-			.unwrap();
-
-		let rm = Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "rm"])
-			.output()
-			.unwrap();
-		assert!(rm.status.success(), "rm failed: {:?}", rm.stderr);
-	}
-
-	#[tokio::test]
-	async fn cli_up_with_build_flag() {
-		if super::podman().await.is_none() {
-			return;
-		}
-		let dir = tempdir().unwrap();
-		let proj = format!("t{}-clbld", std::process::id());
-		fs::write(dir.path().join("Dockerfile"), "FROM alpine:latest\n").unwrap();
-		let compose = dir.path().join("docker-compose.yml");
-		fs::write(
-			&compose,
-			"services:\n  web:\n    build: .\n    command: [\"sleep\", \"infinity\"]\n",
-		)
-		.unwrap();
-
-		let up = Command::new(bin())
-			.args([
-				"-f",
-				compose.to_str().unwrap(),
-				"-p",
-				&proj,
-				"up",
-				"--detach",
-				"--build",
-			])
-			.output()
-			.unwrap();
-		assert!(up.status.success(), "up --build failed: {:?}", up.stderr);
-
-		Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "down"])
-			.output()
-			.unwrap();
-	}
-
-	#[tokio::test]
-	async fn cli_build_subcommand() {
-		if super::podman().await.is_none() {
-			return;
-		}
-		let dir = tempdir().unwrap();
-		fs::write(dir.path().join("Dockerfile"), "FROM alpine:latest\n").unwrap();
-		let compose = dir.path().join("docker-compose.yml");
-		fs::write(
-			&compose,
-			"services:\n  web:\n    build: .\n    command: [\"sleep\", \"infinity\"]\n",
-		)
-		.unwrap();
-
-		let build = Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "build"])
-			.output()
-			.unwrap();
-		assert!(build.status.success(), "build failed: {:?}", build.stderr);
-	}
-
-	#[tokio::test]
-	async fn cli_pause_unpause_subcommands() {
-		if super::podman().await.is_none() {
-			return;
-		}
-		let dir = tempdir().unwrap();
-		let compose = dir.path().join("docker-compose.yml");
-		let proj = format!("t{}-clpu", std::process::id());
-		fs::write(
-			&compose,
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-		)
-		.unwrap();
-
-		Command::new(bin())
-			.args([
-				"-f",
-				compose.to_str().unwrap(),
-				"-p",
-				&proj,
-				"up",
-				"--detach",
-			])
-			.output()
-			.unwrap();
-
-		let pause = Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "pause"])
-			.output()
-			.unwrap();
-		assert!(pause.status.success(), "pause failed: {:?}", pause.stderr);
-
-		let unpause = Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "unpause"])
-			.output()
-			.unwrap();
-		assert!(
-			unpause.status.success(),
-			"unpause failed: {:?}",
-			unpause.stderr
-		);
-
-		Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "down"])
-			.output()
-			.unwrap();
-	}
-
-	#[tokio::test]
-	async fn cli_run_subcommand() {
-		if super::podman().await.is_none() {
-			return;
-		}
-		let dir = tempdir().unwrap();
-		let compose = dir.path().join("docker-compose.yml");
-		let proj = format!("t{}-clrun", std::process::id());
-		fs::write(&compose, "services:\n  job:\n    image: alpine:latest\n").unwrap();
-
-		let run = Command::new(bin())
-			.args([
-				"-f",
-				compose.to_str().unwrap(),
-				"-p",
-				&proj,
-				"run",
-				"job",
-				"echo",
-				"hello",
-			])
-			.output()
-			.unwrap();
-		assert!(run.status.success(), "run failed: {:?}", run.stderr);
-		let stdout = String::from_utf8_lossy(&run.stdout);
-		assert!(stdout.contains("hello"), "expected 'hello' in output");
-	}
-
-	#[tokio::test]
-	async fn cli_run_nonzero_exit_propagates() {
-		if super::podman().await.is_none() {
-			return;
-		}
-		let dir = tempdir().unwrap();
-		let compose = dir.path().join("docker-compose.yml");
-		let proj = format!("t{}-clrxc", std::process::id());
-		fs::write(&compose, "services:\n  job:\n    image: alpine:latest\n").unwrap();
-
-		let run = Command::new(bin())
-			.args([
-				"-f",
-				compose.to_str().unwrap(),
-				"-p",
-				&proj,
-				"run",
-				"job",
-				"false",
-			])
-			.output()
-			.unwrap();
-		assert!(!run.status.success(), "expected non-zero exit from 'false'");
-		assert_eq!(run.status.code(), Some(1));
-	}
-
-	#[tokio::test]
-	async fn cli_top_subcommand() {
-		if super::podman().await.is_none() {
-			return;
-		}
-		let dir = tempdir().unwrap();
-		let compose = dir.path().join("docker-compose.yml");
-		let proj = format!("t{}-cltop", std::process::id());
-		fs::write(
-			&compose,
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-		)
-		.unwrap();
-
-		Command::new(bin())
-			.args([
-				"-f",
-				compose.to_str().unwrap(),
-				"-p",
-				&proj,
-				"up",
-				"--detach",
-			])
-			.output()
-			.unwrap();
-
-		let top = Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "top"])
-			.output()
-			.unwrap();
-		assert!(top.status.success(), "top failed: {:?}", top.stderr);
-
-		Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "down"])
-			.output()
-			.unwrap();
-	}
-
-	#[tokio::test]
-	async fn cli_images_subcommand() {
-		if super::podman().await.is_none() {
-			return;
-		}
-		let dir = tempdir().unwrap();
-		let compose = dir.path().join("docker-compose.yml");
-		let proj = format!("t{}-climg", std::process::id());
-		fs::write(
-			&compose,
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-		)
-		.unwrap();
-
-		Command::new(bin())
-			.args([
-				"-f",
-				compose.to_str().unwrap(),
-				"-p",
-				&proj,
-				"up",
-				"--detach",
-			])
-			.output()
-			.unwrap();
-
-		let images = Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "images"])
-			.output()
-			.unwrap();
-		assert!(
-			images.status.success(),
-			"images failed: {:?}",
-			images.stderr
-		);
-		let stdout = String::from_utf8_lossy(&images.stdout);
-		assert!(
-			stdout.contains("alpine"),
-			"expected 'alpine' in images output"
-		);
-
-		Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "down"])
-			.output()
-			.unwrap();
-	}
-
-	#[tokio::test]
-	async fn cli_port_subcommand() {
-		if super::podman().await.is_none() {
-			return;
-		}
-		let dir = tempdir().unwrap();
-		let compose = dir.path().join("docker-compose.yml");
-		let proj = format!("t{}-clprt", std::process::id());
-		fs::write(
-			&compose,
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n    ports:\n      - \"127.0.0.1:18081:80\"\n",
-		)
-		.unwrap();
-
-		Command::new(bin())
-			.args([
-				"-f",
-				compose.to_str().unwrap(),
-				"-p",
-				&proj,
-				"up",
-				"--detach",
-			])
-			.output()
-			.unwrap();
-
-		let port = Command::new(bin())
-			.args([
-				"-f",
-				compose.to_str().unwrap(),
-				"-p",
-				&proj,
-				"port",
-				"web",
-				"80",
-			])
-			.output()
-			.unwrap();
-		assert!(port.status.success(), "port failed: {:?}", port.stderr);
-
-		Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "down"])
-			.output()
-			.unwrap();
-	}
-
-	#[tokio::test]
-	async fn cli_cp_from_container() {
-		if super::podman().await.is_none() {
-			return;
-		}
-		let dir = tempdir().unwrap();
-		let compose = dir.path().join("docker-compose.yml");
-		let proj = format!("t{}-clcp", std::process::id());
-		fs::write(
-			&compose,
-			"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
-		)
-		.unwrap();
-
-		Command::new(bin())
-			.args([
-				"-f",
-				compose.to_str().unwrap(),
-				"-p",
-				&proj,
-				"up",
-				"--detach",
-			])
-			.output()
-			.unwrap();
-
-		let dst = dir.path().to_str().unwrap();
-		let cp = Command::new(bin())
-			.args([
-				"-f",
-				compose.to_str().unwrap(),
-				"-p",
-				&proj,
-				"cp",
-				"web:/etc/hostname",
-				dst,
-			])
-			.output()
-			.unwrap();
-		assert!(cp.status.success(), "cp failed: {:?}", cp.stderr);
-
-		Command::new(bin())
-			.args(["-f", compose.to_str().unwrap(), "-p", &proj, "down"])
-			.output()
-			.unwrap();
-	}
-}
+#[path = "engine_integration/watch.rs"]
+mod watch_tests;
+
+#[path = "engine_integration/cli1.rs"]
+mod cli1;
+#[path = "engine_integration/cli2.rs"]
+mod cli2;
