+++ tests/engine_integration/cli1.rs
@@ -1,0 +1,336 @@
+//! CLI binary integration tests (covers main.rs).
+use std::fs;
+use std::process::Command;
+use tempfile::tempdir;
+
+use super::*;
+
+#[test]
+fn cli_config_no_podman() {
+	let dir = tempdir().unwrap();
+	let compose = dir.path().join("docker-compose.yml");
+	fs::write(&compose, "services:\n  web:\n    image: alpine:latest\n").unwrap();
+
+	let out = Command::new(bin())
+		.args(["-f", compose.to_str().unwrap(), "config"])
+		.output()
+		.expect("podup binary not found");
+
+	assert!(
+		out.status.success(),
+		"config failed: {}",
+		String::from_utf8_lossy(&out.stderr)
+	);
+	let stdout = String::from_utf8_lossy(&out.stdout);
+	assert!(stdout.contains("alpine"));
+}
+
+#[tokio::test]
+async fn cli_up_and_down_via_binary() {
+	if super::podman().await.is_none() {
+		return;
+	}
+	let dir = tempdir().unwrap();
+	let compose = dir.path().join("docker-compose.yml");
+	let pid = std::process::id();
+	let proj = format!("t{}-cli", pid);
+	fs::write(
+		&compose,
+		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
+	)
+	.unwrap();
+
+	let up = Command::new(bin())
+		.args([
+			"-f",
+			compose.to_str().unwrap(),
+			"-p",
+			&proj,
+			"up",
+			"--detach",
+		])
+		.output()
+		.unwrap();
+	assert!(
+		up.status.success(),
+		"up failed: {}",
+		String::from_utf8_lossy(&up.stderr)
+	);
+
+	let down = Command::new(bin())
+		.args(["-f", compose.to_str().unwrap(), "-p", &proj, "down"])
+		.output()
+		.unwrap();
+	assert!(
+		down.status.success(),
+		"down failed: {}",
+		String::from_utf8_lossy(&down.stderr)
+	);
+}
+
+#[tokio::test]
+async fn cli_ps_subcommand() {
+	if super::podman().await.is_none() {
+		return;
+	}
+	let dir = tempdir().unwrap();
+	let compose = dir.path().join("docker-compose.yml");
+	let proj = format!("t{}-clps", std::process::id());
+	fs::write(
+		&compose,
+		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
+	)
+	.unwrap();
+
+	Command::new(bin())
+		.args([
+			"-f",
+			compose.to_str().unwrap(),
+			"-p",
+			&proj,
+			"up",
+			"--detach",
+		])
+		.output()
+		.unwrap();
+
+	let ps = Command::new(bin())
+		.args(["-f", compose.to_str().unwrap(), "-p", &proj, "ps"])
+		.output()
+		.unwrap();
+	assert!(ps.status.success());
+
+	Command::new(bin())
+		.args(["-f", compose.to_str().unwrap(), "-p", &proj, "down"])
+		.output()
+		.unwrap();
+}
+
+#[tokio::test]
+async fn cli_logs_subcommand() {
+	if super::podman().await.is_none() {
+		return;
+	}
+	let dir = tempdir().unwrap();
+	let compose = dir.path().join("docker-compose.yml");
+	let proj = format!("t{}-cllg", std::process::id());
+	fs::write(
+		&compose,
+		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
+	)
+	.unwrap();
+
+	Command::new(bin())
+		.args([
+			"-f",
+			compose.to_str().unwrap(),
+			"-p",
+			&proj,
+			"up",
+			"--detach",
+		])
+		.output()
+		.unwrap();
+
+	let logs = Command::new(bin())
+		.args(["-f", compose.to_str().unwrap(), "-p", &proj, "logs"])
+		.output()
+		.unwrap();
+	assert!(logs.status.success());
+
+	Command::new(bin())
+		.args(["-f", compose.to_str().unwrap(), "-p", &proj, "down"])
+		.output()
+		.unwrap();
+}
+
+#[tokio::test]
+async fn cli_exec_subcommand() {
+	if super::podman().await.is_none() {
+		return;
+	}
+	let dir = tempdir().unwrap();
+	let compose = dir.path().join("docker-compose.yml");
+	let proj = format!("t{}-clex", std::process::id());
+	fs::write(
+		&compose,
+		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
+	)
+	.unwrap();
+
+	Command::new(bin())
+		.args([
+			"-f",
+			compose.to_str().unwrap(),
+			"-p",
+			&proj,
+			"up",
+			"--detach",
+		])
+		.output()
+		.unwrap();
+
+	let exec = Command::new(bin())
+		.args([
+			"-f",
+			compose.to_str().unwrap(),
+			"-p",
+			&proj,
+			"exec",
+			"web",
+			"echo",
+			"cli-exec",
+		])
+		.output()
+		.unwrap();
+	assert!(exec.status.success());
+
+	Command::new(bin())
+		.args(["-f", compose.to_str().unwrap(), "-p", &proj, "down"])
+		.output()
+		.unwrap();
+}
+
+#[tokio::test]
+async fn cli_restart_subcommand() {
+	if super::podman().await.is_none() {
+		return;
+	}
+	let dir = tempdir().unwrap();
+	let compose = dir.path().join("docker-compose.yml");
+	let proj = format!("t{}-clrs", std::process::id());
+	fs::write(
+		&compose,
+		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
+	)
+	.unwrap();
+
+	Command::new(bin())
+		.args([
+			"-f",
+			compose.to_str().unwrap(),
+			"-p",
+			&proj,
+			"up",
+			"--detach",
+		])
+		.output()
+		.unwrap();
+
+	let restart = Command::new(bin())
+		.args(["-f", compose.to_str().unwrap(), "-p", &proj, "restart"])
+		.output()
+		.unwrap();
+	assert!(restart.status.success());
+
+	Command::new(bin())
+		.args(["-f", compose.to_str().unwrap(), "-p", &proj, "down"])
+		.output()
+		.unwrap();
+}
+
+#[tokio::test]
+async fn cli_pull_subcommand() {
+	if super::podman().await.is_none() {
+		return;
+	}
+	let dir = tempdir().unwrap();
+	let compose = dir.path().join("docker-compose.yml");
+	fs::write(&compose, "services:\n  web:\n    image: alpine:latest\n").unwrap();
+
+	let pull = Command::new(bin())
+		.args(["-f", compose.to_str().unwrap(), "pull"])
+		.output()
+		.unwrap();
+	assert!(pull.status.success());
+}
+
+#[tokio::test]
+async fn cli_stop_and_start_subcommands() {
+	if super::podman().await.is_none() {
+		return;
+	}
+	let dir = tempdir().unwrap();
+	let compose = dir.path().join("docker-compose.yml");
+	let proj = format!("t{}-clss", std::process::id());
+	fs::write(
+		&compose,
+		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
+	)
+	.unwrap();
+
+	Command::new(bin())
+		.args([
+			"-f",
+			compose.to_str().unwrap(),
+			"-p",
+			&proj,
+			"up",
+			"--detach",
+		])
+		.output()
+		.unwrap();
+
+	let stop = Command::new(bin())
+		.args(["-f", compose.to_str().unwrap(), "-p", &proj, "stop"])
+		.output()
+		.unwrap();
+	assert!(stop.status.success(), "stop failed: {:?}", stop.stderr);
+
+	let start = Command::new(bin())
+		.args(["-f", compose.to_str().unwrap(), "-p", &proj, "start"])
+		.output()
+		.unwrap();
+	assert!(start.status.success(), "start failed: {:?}", start.stderr);
+
+	Command::new(bin())
+		.args(["-f", compose.to_str().unwrap(), "-p", &proj, "down"])
+		.output()
+		.unwrap();
+}
+
+#[tokio::test]
+async fn cli_kill_subcommand() {
+	if super::podman().await.is_none() {
+		return;
+	}
+	let dir = tempdir().unwrap();
+	let compose = dir.path().join("docker-compose.yml");
+	let proj = format!("t{}-clkl", std::process::id());
+	fs::write(
+		&compose,
+		"services:\n  web:\n    image: alpine:latest\n    command: [\"sleep\", \"infinity\"]\n",
+	)
+	.unwrap();
+
+	Command::new(bin())
+		.args([
+			"-f",
+			compose.to_str().unwrap(),
+			"-p",
+			&proj,
+			"up",
+			"--detach",
+		])
+		.output()
+		.unwrap();
+
+	let kill = Command::new(bin())
+		.args([
+			"-f",
+			compose.to_str().unwrap(),
+			"-p",
+			&proj,
+			"kill",
+			"--signal",
+			"SIGTERM",
+		])
+		.output()
+		.unwrap();
+	assert!(kill.status.success(), "kill failed: {:?}", kill.stderr);
+
+	Command::new(bin())
+		.args(["-f", compose.to_str().unwrap(), "-p", &proj, "down"])
+		.output()
+		.unwrap();
+}
